move system_time serde module to crate level to fix layering inversion
Provider types referenced crate::public::proto_json::system_time for
timestamp serde attributes, which made the provider surface depend on
the public client module. The system_time module is now at
crate::serde_time, a crate-private module with no dependency on
crate::public. The old re-export from proto_json.rs is removed since
nothing references the old path after regeneration.
